Added hook functionality
